Move hooks to Hooks class
